Move the last sentinel-1 specific code in apps to sentinel1
